backend

+++ s3-service/src/main/java/com/aws/s3service/S3Service.java
@@ -1,5 +1,7 @@
 package com.aws.s3service;
 
+import com.aws.s3service.exception.NotFoundException;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import software.amazon.awssdk.core.ResponseBytes;
@@ -8,8 +10,8 @@ import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.model.*;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
+@Slf4j
 @Service
 public class S3Service {
 
@@ -23,27 +25,36 @@ public class S3Service {
     }
 
     public String upload(String key, byte[] content) {
+        log.info("Subiendo archivo: {} al bucket: {}", key, bucket);
         s3.putObject(PutObjectRequest.builder().bucket(bucket).key(key).build(),
                 RequestBody.fromBytes(content));
         return "Archivo subido: " + key;
     }
 
     public byte[] download(String key) {
+        log.info("Descargando archivo: {} desde el bucket: {}", key, bucket);
         ResponseBytes<GetObjectResponse> objectBytes = s3.getObjectAsBytes(
-                GetObjectRequest.builder().bucket(bucket).key(key).build()
-        );
+                GetObjectRequest.builder().bucket(bucket).key(key).build());
         return objectBytes.asByteArray();
     }
 
     public List<String> list() {
-        return s3.listObjectsV2(ListObjectsV2Request.builder().bucket(bucket).build())
+        log.info("Listando archivos en el bucket: {}", bucket);
+        List<String> res = s3.listObjectsV2(ListObjectsV2Request.builder().bucket(bucket).build())
                 .contents()
                 .stream()
                 .map(S3Object::key)
-                .collect(Collectors.toList());
+                .toList();
+        log.info("Archivos encontrados: {}", res);
+        return res;
     }
 
     public String delete(String key) {
+        List<String> files = list();
+        if (!files.contains(key)) {
+            log.warn("Archivo no encontrado: {}", key);
+            throw new NotFoundException("Archivo no encontrado: " + key);
+        }
         s3.deleteObject(DeleteObjectRequest.builder().bucket(bucket).key(key).build());
         return "Archivo eliminado: " + key;
     }
